Implementation/requirements.txt
oyaml
comet-ml
pillow
numpy
torch
torchvision
opendatasets